backend/requirements.txt
# Core Framework (Python 3.9 compatible versions)
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.1
python-dotenv==1.0.1

# HTTP Client
httpx==0.28.0
requests==2.32.3

# Cryptography for Kalshi API signing
cryptography==44.0.0
pyjwt==2.10.1

# Data Processing (Python 3.9 compatible)
numpy==1.26.4
pandas==2.0.3
scipy==1.11.4

# Database (Supabase)
supabase==2.10.0
psycopg2-binary==2.9.10

# Async Support
aiofiles==24.1.0

# Logging & Monitoring
python-json-logger==3.2.1
colorlog==6.9.0

# Configuration
pydantic==2.10.3
pydantic-settings==2.7.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0
